requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
fastapi==0.115.0
feedparser==6.0.11
httpx==0.27.2
lxml==5.3.0
numpy==1.26.4
pandas==2.2.3
pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
sentence-transformers==3.1.1
spacy==3.8.2
torch>=2.3.0
torchvision>=0.18.0
transformers==4.45.2
uvicorn==0.30.6
